requirements.txt
app-store-scraper==0.3.5
beautifulsoup4==4.10.0
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
expertai-nlapi==2.3.1
Flask==2.0.1
google-play-scraper==1.0.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
Pillow==8.3.2
plotly==5.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
regex==2021.9.30
requests==2.23.0
requests-oauthlib==1.3.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.2.1
tenacity==8.0.1
textblob==0.15.3
tqdm==4.62.3
tweepy==4.0.1
urllib3==1.25.11
Werkzeug==2.0.1
gensim
Flask-Session
pyarrow
